graphql-gateway/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
ariadne
python-dotenv==1.0.0
mysql-connector-python==8.1.0
Flask-Cors==4.0.0
pytest==7.4.0
pytest-mock==3.11.1
httpx
requests